Add files via upload
The project is based on AI, DS and ML. In this project I have made and trained models using R programming language. This project was made to detect the default of credit card clients for the coming month. This was quite interesting to made and I learned a lot of new things making this project. I am uploading this project for the sake of help to the newbies.